feat: 체크박스 컴포넌트화
