+++ src/components/ui/Checkbox/Checkbox.tsx
@@ -1,6 +1,6 @@
-import { Ionicons } from '@expo/vector-icons';
 import React from 'react';
-import { Text, TouchableOpacity, View } from 'react-native';
+import { Text, TouchableOpacity } from 'react-native';
+import CheckboxBox from './CheckboxBox';
 
 interface CheckboxProps {
   checked: boolean;
@@ -17,26 +17,8 @@ export default function Checkbox({
   size = 'md',
   disabled = false,
 }: CheckboxProps) {
-  const sizeConfig = {
-    sm: { box: 'w-3.5 h-3.5', icon: 10 },
-    md: { box: 'w-4 h-4', icon: 12 },
-    lg: { box: 'w-5 h-5', icon: 14 },
-  };
-
-  const { box, icon } = sizeConfig[size];
-
-  const CheckboxBox = (
-    <View
-      className={`${box} rounded border items-center justify-center ${
-        checked ? 'bg-primary-blue border-primary-blue' : 'border-gray-300'
-      }`}
-    >
-      {checked && <Ionicons name="checkmark" size={icon} color="white" />}
-    </View>
-  );
-
   if (!label) {
-    return CheckboxBox;
+    return <CheckboxBox checked={checked} size={size} />;
   }
 
   return (
@@ -46,7 +28,7 @@ export default function Checkbox({
       className="flex-row items-center gap-2"
       activeOpacity={0.7}
     >
-      {CheckboxBox}
+      <CheckboxBox checked={checked} size={size} />
       <Text
         className={`text-sm ${disabled ? 'text-gray-400' : 'text-gray-900'}`}
       >
