+++ src/components/ui/Checkbox/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Checkbox';
+export { default as CheckboxBox } from './CheckboxBox';
